Added the 'Add item' screen/dialog on Presets screen (with validation)
Added the 'tap' handling on SavedFilesViewModel.cs, now it fires an event
that is handled by the main pivot view - activates the 'current' view model and switches
the pivot page.

Changed the Frequency type from double to int

+++ PureToneAudiometer/Audio/IOscillator.cs
@@ -3,7 +3,7 @@
     public interface IOscillator
     {
         int Attenuation { get; set; }
-        double Frequency { get; set; }
+        int Frequency { get; set; }
         short GetValue();
         void Reset();
     }
